module8-task2 добавил логику выбора случайного города
