+++ src/helpers/getRandomCity.tsx
@@ -1,0 +1,1 @@
+export const getRandomCity = (array: string[]): string => array[Math.floor(Math.random() * array.length)];
