+++ src/pages/favorites.tsx
@@ -2,35 +2,20 @@ import Container from '../components/container';
 import FavoriteLocation from '../components/favorite-location';
 import {useAppSelector} from '../hooks';
 import {getFavoriteOffers} from '../store/slices/favorite/selectors';
-import {Offer} from '../types/offer';
 import EmptyFavoriteOffers from '../components/empty-favorite-offers';
 import classNames from 'classnames';
-
-type FilteredDataType = {
-  city: string;
-  offers: Offer[];
-};
-
-const createMap = (data: Offer[]): FilteredDataType[] | [] => {
-  const map = new Map<string, Offer[]>();
-  const res: FilteredDataType[] = [];
-  data.forEach((el, i, arr) => {
-    map.set(el.city.name, [...(map.get(el.city.name) || []), ...[arr[i]]]);
-  });
-
-  map.forEach((value, key) => {
-    res.push({city: key, offers: value});
-  });
-
-  return res;
-};
+import {useMemo} from 'react';
+import {preparedFavoriteData} from '../helpers/preparedFavoriteData';
 
 export default function Favorites() {
   const offers = useAppSelector(getFavoriteOffers);
-  const preparedData = createMap(offers);
+  const preparedData = useMemo(() => preparedFavoriteData(offers), [offers]);
+
   return (
     <Container
-      extraClass='page--favorites-empty'
+      extraClass={classNames({
+        'page--favorites-empty': offers.length === 0,
+      })}
       classMain={classNames({
         'page__main--favorites': true,
         'page__main--favorites-empty': offers.length === 0,
