+++ src/pages/login.tsx
@@ -1,11 +1,18 @@
 import Container from '../components/container';
-import {useRef, FormEvent} from 'react';
+import {useRef, useMemo, FormEvent} from 'react';
 import {useAppDispatch} from '../hooks';
+import {useAppSelector} from '../hooks';
 import {loginAction} from '../store/thunks/user';
+import {getCitiesNames} from '../store/slices/cities/selectors';
+import {getRandomCity} from '../helpers/getRandomCity';
+import {setCity} from '../store/slices/cities/cities';
+import {Link} from 'react-router-dom';
 
 export default function Login() {
   const loginRef = useRef<HTMLInputElement | null>(null);
   const passwordRef = useRef<HTMLInputElement | null>(null);
+  const citiesNames = useAppSelector(getCitiesNames);
+  const randomCity = useMemo(() => getRandomCity(citiesNames), []);
 
   const dispatch = useAppDispatch();
 
@@ -18,6 +25,10 @@ export default function Login() {
         password: passwordRef.current?.value
       }));
     }
+  };
+
+  const randomCityHandler = (city: string) => {
+    dispatch(setCity(city));
   };
 
   return (
@@ -51,9 +62,13 @@ export default function Login() {
         </section>
         <section className="locations locations--login locations--current">
           <div className="locations__item">
-            <a className="locations__item-link" href="#">
-              <span>Amsterdam</span>
-            </a>
+            <Link
+              className="locations__item-link"
+              to="/"
+              onClick={() => randomCityHandler(randomCity)}
+            >
+              <span>{randomCity}</span>
+            </Link>
           </div>
         </section>
       </div>
